+++ cjio/cityjson.py
@@ -1,13 +1,11 @@
 
 import os
-import sys
 import re
 import shutil
 
 import json
 import collections
 import jsonref
-import urllib
 from pkg_resources import resource_filename
 from pkg_resources import resource_listdir
 import copy
@@ -16,18 +14,23 @@ from io import StringIO
 import numpy as np
 import pyproj
 from sys import platform
+from click import progressbar
 
 MODULE_EARCUT_AVAILABLE = True
+MODULE_PANDAS_AVAILABLE = True
+
 try:
     import mapbox_earcut
 except ImportError as e:
     MODULE_EARCUT_AVAILABLE = False
+try:
+    import pandas
+except ImportError as e:
+    MODULE_PANDAS_AVAILABLE = False
 
-from cjio import validation
-from cjio import subset
-from cjio import geom_help
-from cjio import errors
+from cjio import validation, subset, geom_help, convert, models
 from cjio.errors import InvalidOperation
+from cjio.utils import print_cmd_warning
 
 
 CITYJSON_VERSIONS_SUPPORTED = ['0.6', '0.8', '0.9', '1.0']
@@ -47,6 +50,73 @@ TOPLEVEL = ('Building',
             'Tunnel',
             'WaterBody')
 
+def load(path, transform:bool=False):
+    """Load a CityJSON file for working with it though the API
+
+    :param path: Absolute path to a CityJSON file
+    :param transform: Apply the coordinate transformation to the vertices (if applicable)
+    :return: A CityJSON object
+    """
+    with open(path, 'r') as fin:
+        try:
+            cm = CityJSON(file=fin)
+        except OSError as e:
+            raise FileNotFoundError
+    cm.cityobjects = dict()
+    if 'transform' in cm.j:
+        cm.transform = cm.j['transform']
+    else:
+        cm.transform = None
+    if transform:
+        do_transform = cm.transform
+        del cm.j['transform']
+    else:
+        do_transform = None
+    for co_id, co in cm.j['CityObjects'].items():
+        # TODO BD: do some verification here
+        children = co['children'] if 'children' in co else None
+        parents = co['parents'] if 'parents' in co else None
+        attributes = co['attributes'] if 'attributes' in co else None
+        geometry = []
+        for geom in co['geometry']:
+            semantics = geom['semantics'] if 'semantics' in geom else None
+            geometry.append(
+                models.Geometry(
+                    type=geom['type'],
+                    lod=geom['lod'],
+                    boundaries=geom['boundaries'],
+                    semantics_obj=semantics,
+                    vertices=cm.j['vertices'],
+                    transform=do_transform
+                )
+            )
+        cm.cityobjects[co_id] = models.CityObject(
+            id=co_id,
+            type=co['type'],
+            attributes=attributes,
+            children=children,
+            parents=parents,
+            geometry=geometry
+        )
+    return cm
+
+def save(citymodel, path: str):
+    """Save a city model to a CityJSON file
+
+    :param citymodel: A CityJSON object
+    :param path: Absolute path to a CityJSON file
+    """
+    cityobjects, vertex_lookup = citymodel.reference_geometry()
+    citymodel.j['vertices'] = list(vertex_lookup.keys())
+    citymodel.j['CityObjects'] = cityobjects
+    citymodel.remove_duplicate_vertices()
+    citymodel.remove_orphan_vertices()
+    try:
+        with open(path, 'w') as fout:
+            json_str = json.dumps(citymodel.j, indent=2)
+            fout.write(json_str)
+    except IOError as e:
+        raise IOError('Invalid output file: %s \n%s' % (path, e))
 
 def reader(file, ignore_duplicate_keys=False):
     return CityJSON(file=file, ignore_duplicate_keys=ignore_duplicate_keys)
@@ -68,7 +138,7 @@ def off2cj(file):
         lstFaces.append(list(map(int, file.readline().split()[1:])))
     cm = {}
     cm["type"] = "CityJSON"
-    cm["version"] = "0.6"
+    cm["version"] = CITYJSON_VERSIONS_SUPPORTED[-1]
     cm["CityObjects"] = {}
     cm["vertices"] = []
     for v in lstVertices:
@@ -105,7 +175,7 @@ def poly2cj(file):
         lstFaces.append(face)
     cm = {}
     cm["type"] = "CityJSON"
-    cm["version"] = "0.6"
+    cm["version"] = CITYJSON_VERSIONS_SUPPORTED[-1]
     cm["CityObjects"] = {}
     cm["vertices"] = []
     for v in lstVertices:
@@ -128,19 +198,86 @@ class CityJSON:
         if file is not None:
             self.read(file, ignore_duplicate_keys)
             self.path = os.path.abspath(file.name)
+            self.cityobjects = None
         elif j is not None:
             self.j = j
+            self.cityobjects = None
         else: #-- create an empty one
             self.j = {}
             self.j["type"] = "CityJSON"
             self.j["version"] = CITYJSON_VERSIONS_SUPPORTED[-1]
             self.j["CityObjects"] = {}
             self.j["vertices"] = []
+            self.cityobjects = None
 
 
     def __repr__(self):
         return self.get_info()
 
+
+    ##-- API functions
+    # TODO BD: refactor this whole CityJSON class
+    def get_cityobjects(self, type=None, id=None):
+        """Return a subset of CityObjects
+
+        :param type: CityObject type. If a list of types are given, then all types in the list are returned.
+        :param id: CityObject ID. If a list of IDs are given, then all objects matching the IDs in the list are returned.
+        """
+        if type is None and id is None:
+            return self.cityobjects
+        elif (type is not None) and (id is not None):
+            raise AttributeError("Please provide either 'type' or 'id'")
+        elif type is not None:
+            if isinstance(type, str):
+                type_list = [type.lower()]
+            elif isinstance(type, list) and isinstance(type[0], str):
+                type_list = [t.lower() for t in type]
+            else:
+                raise TypeError("'type' must be a string or list of strings")
+            return {i:co for i,co in self.cityobjects.items() if co.type.lower() in type_list}
+        elif id is not None:
+            if isinstance(id, str):
+                id_list = [id]
+            elif isinstance(id, list) and isinstance(id[0], str):
+                id_list = id
+            else:
+                raise TypeError("'id' must be a string or list of strings")
+            return {i:co for i,co in self.cityobjects.items() if co.id in id_list}
+
+
+    def set_cityobjects(self, cityobjects):
+        """Creates or updates CityObjects
+
+        .. note:: If a CityObject with the same ID already exists in the model, it will be overwritten
+
+        :param cityobjects: Dictionary of CityObjects, where keys are the CityObject IDs. Same structure as returned by get_cityobjects()
+        """
+        for co_id, co in cityobjects.items():
+            self.cityobjects[co_id] = co
+
+
+    def to_dataframe(self):
+        """Converts the city model to a Pandas data frame where fields are CityObject attributes"""
+        if not MODULE_PANDAS_AVAILABLE:
+            raise ModuleNotFoundError("pandas is not available, please install it")
+        return pandas.DataFrame([co.attributes for co_id,co in self.cityobjects.items()],
+                                index=list(self.cityobjects.keys()))
+
+
+    def reference_geometry(self):
+        """Build a coordinate list and index the vertices"""
+        cityobjects = dict()
+        vertex_lookup = dict()
+        vertex_idx = 0
+        for co_id, co in self.cityobjects.items():
+            j_co = co.to_json()
+            geometry, vertex_lookup, vertex_idx = co.build_index(vertex_lookup, vertex_idx)
+            j_co['geometry'] = geometry
+            cityobjects[co_id] = j_co
+        return (cityobjects, vertex_lookup)
+
+
+    ##-- end API functions
 
     def get_version(self):
         return self.j["version"]
@@ -153,7 +290,11 @@ class CityJSON:
             return self.j["metadata"]["crs"]["epsg"]
         elif "referenceSystem" in self.j["metadata"]:
             s = self.j["metadata"]["referenceSystem"]
-            return int(s[s.find("::")+2:])
+            if "epsg" in s.lower():
+                return int(s[s.find("::")+2:])
+            else:
+                print_cmd_warning("Only EPSG codes are supported in the URN. CRS is set to undefined.")
+                return None
         else:
             return None
 
@@ -348,17 +489,24 @@ class CityJSON:
                 if (isValid == False):
                     es += errs
                     return (False, False, es, [])
+
         #-- 2. schema for Extensions
         if "extensions" in self.j:
             b, es = self.validate_extensions(folder_schemas)
             if b == False:
                 return (b, True, es, [])
+        else:
+            #-- check that there are no +CityObject that do not have a schema
+            #-- (used if the file has no "extensions" property while there are +Pand for instance)
+            for theid in self.j["CityObjects"]:
+                if ( (self.j["CityObjects"][theid]["type"][0] == "+") ):
+                    s = "ERROR:   CityObject " + self.j["CityObjects"][theid]["type"] + " doesn't have a schema."
+                    es.append(s)
+                    return (False, False, es, [])
 
-
-        #-- 3. ERRORS
+        #-- 3. Internal consistency validation 
         print ('-- Validating the internal consistency of the file (see docs for list)')
         isValid = True
-
         if float(self.j["version"]) == 0.6:
             b, errs = validation.building_parts(self.j) 
             if b == False:
@@ -583,8 +731,7 @@ class CityJSON:
                     re.add(child)
             if "parents" in self.j['CityObjects'][theid]:
                 for each in self.j['CityObjects'][theid]['parents']:
-                    re.add(self.j['CityObjects'][each])
-
+                    re.add(each)
         
         for each in re:
             cm2.j["CityObjects"][each] = self.j["CityObjects"][each]
@@ -639,6 +786,8 @@ class CityJSON:
         cm2 = CityJSON()
         cm2.j["version"] = self.j["version"]
         cm2.path = self.path
+        if "extensions" in self.j:
+            cm2.j["extensions"] = self.j["extensions"]
         if "transform" in self.j:
             cm2.j["transform"] = self.j["transform"]
         #-- copy selected CO to the j2
@@ -863,7 +1012,7 @@ class CityJSON:
         return total
 
 
-    def get_info(self):
+    def get_info(self, long=False):
         info = collections.OrderedDict()
         info["cityjson_version"] = self.get_version()
         info["epsg"] = self.get_epsg()
@@ -873,24 +1022,43 @@ class CityJSON:
             for i in self.j["extensions"]:
                 d.add(i)
             info["extensions"] = sorted(list(d))
+        info["transform/compressed"] = "transform" in self.j
         info["cityobjects_total"] = self.number_city_objects()
         d = set()
         for key in self.j["CityObjects"]:
             d.add(self.j['CityObjects'][key]['type'])
         info["cityobjects_present"] = sorted(list(d))
-        info["vertices_total"] = len(self.j["vertices"])
-        info["transform/compressed"] = "transform" in self.j
-        d.clear()
-        for key in self.j["CityObjects"]:
-            for geom in self.j['CityObjects'][key]['geometry']:
-                d.add(geom["type"])
-        info["geom_primitives_present"] = list(d)
         if 'appearance' in self.j:
             info["materials"] = 'materials' in self.j['appearance']
             info["textures"] = 'textures' in self.j['appearance']
         else:
             info["materials"] = False
             info["textures"] =  False
+        if long == False:
+            return json.dumps(info, indent=2)    
+        #-- all/long version
+        info["vertices_total"] = len(self.j["vertices"])
+        d.clear()
+        lod = set()
+        sem_srf = set()
+        co_attributes = set()
+        for key in self.j["CityObjects"]:
+            if 'attributes' in self.j['CityObjects'][key]:
+                for attr in self.j['CityObjects'][key]['attributes'].keys():
+                    co_attributes.add(attr)
+            for geom in self.j['CityObjects'][key]['geometry']:
+                d.add(geom["type"])
+                if "lod" in geom:
+                    lod.add(geom["lod"])
+                else: #-- it's a geometry-template
+                    lod.add(self.j["geometry-templates"]["templates"][geom["template"]]["lod"])
+                if "semantics" in geom:
+                    for srf in geom["semantics"]["surfaces"]:
+                        sem_srf.add(srf["type"])
+        info["geom_primitives_present"] = list(d)
+        info["level_of_detail"] = list(lod)
+        info["semantics_surfaces_present"] = list(sem_srf)
+        info["cityobject_attributes"] = list(co_attributes)
         return json.dumps(info, indent=2)
 
 
@@ -1042,22 +1210,32 @@ class CityJSON:
         for cm in lsCMs:
             #-- decompress 
             cm.decompress()
-            #-- add each CityObjects
-            coadded = 0
-            for theid in cm.j["CityObjects"]:
-                if theid in self.j["CityObjects"]:
-                    print ("ERROR: CityObject #", theid, "already present. Skipped.")
-                else:
-                    self.j["CityObjects"][theid] = cm.j["CityObjects"][theid]
-                    coadded += 1
-            if coadded == 0:
-                continue
-            #-- add the vertices + update the geom indices
             offset = len(self.j["vertices"])
             self.j["vertices"] += cm.j["vertices"]
+            #-- add each CityObjects
             for theid in cm.j["CityObjects"]:
-                for g in cm.j['CityObjects'][theid]['geometry']:
-                    update_geom_indices(g["boundaries"], offset)
+                if theid in self.j["CityObjects"]:
+                    #-- merge attributes if not present (based on the property name only)
+                    for a in cm.j["CityObjects"][theid]["attributes"]:
+                        if a not in self.j["CityObjects"][theid]["attributes"]:
+                            self.j["CityObjects"][theid]["attributes"][a] = cm.j["CityObjects"][theid]["attributes"][a]
+                    #-- merge geoms if not present (based on LoD only)
+                    for g in cm.j['CityObjects'][theid]['geometry']:
+                        thelod = str(g["lod"])
+                        # print ("-->", thelod)
+                        b = False
+                        for g2 in self.j['CityObjects'][theid]['geometry']:
+                            if g2["lod"] == thelod:
+                                b = True
+                                break
+                        if b == False:
+                            self.j['CityObjects'][theid]['geometry'].append(g)
+                            update_geom_indices(self.j['CityObjects'][theid]['geometry'][-1]["boundaries"], offset)    
+                else:
+                    #-- copy the CO
+                    self.j["CityObjects"][theid] = cm.j["CityObjects"][theid]
+                    for g in self.j['CityObjects'][theid]['geometry']:
+                        update_geom_indices(g["boundaries"], offset)
             #-- templates
             if "geometry-templates" in cm.j:
                 if "geometry-templates" in self.j:
@@ -1125,8 +1303,8 @@ class CityJSON:
                         if 'texture' in g:
                             for m in g['texture']:
                                 update_texture_indices(g['texture'][m]['values'], toffset, voffset)
-        # self.remove_duplicate_vertices()
-        # self.remove_orphan_vertices()
+        self.remove_duplicate_vertices()
+        self.remove_orphan_vertices()
         return True
 
 
@@ -1254,7 +1432,21 @@ class CityJSON:
         return (result.reshape(-1, 3), True)
 
 
+    def export2b3dm(self):
+        glb = convert.to_glb(self.j)
+        b3dm = convert.to_b3dm(self, glb)
+        return b3dm
+
+
+    def export2gltf(self):
+        # TODO B: probably no need to double wrap this to_gltf(), but its long, and
+        # the current cityjson.py is long already
+        glb = convert.to_glb(self.j)
+        return glb
+
+
     def export2obj(self):
+        self.decompress()
         out = StringIO()
         #-- write vertices
         for v in self.j['vertices']:
@@ -1300,11 +1492,12 @@ class CityJSON:
             wascompressed = True
         p1 = pyproj.Proj(init='epsg:%d' % (self.get_epsg()))
         p2 = pyproj.Proj(init='epsg:%d' % (epsg))
-        for v in self.j['vertices']:
-            x, y, z = pyproj.transform(p1, p2, v[0], v[1], v[2])
-            v[0] = x
-            v[1] = y
-            v[2] = z
+        with progressbar(self.j['vertices']) as vertices:
+            for v in vertices:
+                x, y, z = pyproj.transform(p1, p2, v[0], v[1], v[2])
+                v[0] = x
+                v[1] = y
+                v[2] = z
         self.set_epsg(epsg)
         if wascompressed == True:
             self.compress()